Fix clippy lints in sysbox

+++ src/sys/tools/sysbox/src/commands/ps.rs
@@ -40,7 +40,7 @@ pub fn do_command(args: &[String]) {
         Ok(cnt) => cnt,
         Err(err) => {
             eprintln!("PS failed.");
-            std::process::exit(err as u16 as i32);
+            std::process::exit(err as i32);
         }
     };
